+++ front/components/Header.vue
@@ -25,86 +25,18 @@
           </div>
 
           <div v-else>
-            <v-dialog
-              transition="fab-transition"
-              v-model="dialog"
-              persistent
-              max-width="600px"
+            <v-btn to ="/loggedIn">
+              logout
+            </v-btn>
+            <v-btn
+              color="primary" 
+              dark 
+              large
+              to ="/loginForm"
             >
-              <template v-slot:activator="{ on, attrs }">
-                <v-btn
-                  color="primary" 
-                  dark 
-                  v-bind="attrs" 
-                  v-on="on"
-                  large
-                >
-                  ログイン
-                  <v-icon>mdi-account-arrow-left-outline</v-icon>
-                </v-btn>
-              </template>
-
-              <v-card>
-                <v-card-title>
-                  <div class="text-h5">ログイン</div>
-                </v-card-title>
-                <v-card-text>
-                  <v-container>
-                    <v-row>
-                      <!-- <v-col cols="12" sm="6" md="4">
-                        <v-text-field
-                          v-model ="userId"
-                          label="ユーザー名*"
-                          required
-                        ></v-text-field>
-                      </v-col> -->
-                      <v-col cols="12">
-                        <v-text-field
-                          v-model ="email"
-                          label="メールアドレス*"
-                          required
-                        ></v-text-field>
-                      </v-col>
-                      <v-col cols="12">
-                        <v-text-field
-                          v-model ="password"
-                          label="パスワード*"
-                          type="password"
-                          required
-                        ></v-text-field>
-                      </v-col>
-                    </v-row>
-                  </v-container>
-                  <small>*は必須項目です</small>
-                </v-card-text>
-                <div>
-                  <v-card-actions class="justify-center">
-                    <v-spacer></v-spacer>
-                    <v-btn 
-                      color="blue darken-1" 
-                      text 
-                      @click="dialog = false"
-                    >
-                      閉じる
-                    </v-btn>
-                    <v-btn 
-                      
-                      color="blue darken-1" 
-                      text 
-                      @click="login"
-                    >
-                      ログイン
-                    </v-btn>
-                  </v-card-actions>
-                </div>
-                <v-card-text>
-                  <div>
-                    ユーザー登録はお済みですか？
-                    <NuxtLink to="/user/create">ー新規登録</NuxtLink>
-                  </div>
-                </v-card-text>
-              </v-card>
-            </v-dialog>
+              ログイン
+              <v-icon>mdi-account-arrow-left-outline</v-icon>
+            </v-btn>
 
             <v-btn 
               color="" 
@@ -126,42 +58,18 @@
 import Vuex from 'vuex';
 
 export default {
-  data: () => ({
-    //loginwindow
-    dialog: false,
-    // userId: '',
-    email: '',
-    password: '',
-  }),
 
   computed: {
     user () {
       return this.$auth.user
     }
   },
-
-  methods: {
-    async login () {
-      try {
-        const res = await this.$auth.loginWith('local', {
-          data: {
-            // user_id: this.userId,
-            email: this.email,
-            password: this.password
-          }
-        })
-        console.log(res)
-        this.$router.push('/')
-      }catch (err){
-        console.log(err)
-      }
-    },
-
+  method: {
     logout() {
       this.$auth.logout();
     },
-
   },
+
 }
 
 </script>
