+++ front/plugins/axios/modules/user.js
@@ -1,0 +1,11 @@
+import { axios } from '../index.js';
+
+export default {
+  getUser(id) {
+    return axios.$get(`users/${id}`)
+  },
+
+  getUsers() {
+    return axios.$get(`users`)
+  }
+}
