+++ front/plugins/localStrage.js
@@ -1,9 +1,9 @@
-import createPersistedState from 'vuex-persistedstate'
+import createPersistedState from 'vuex-persistedstate';
 
 export default ({ store }) => {
   window.onNuxtReady(() => {
     createPersistedState({
       key: 'persisted-key'
     })(store)
-  })
+  });
 }
