+++ front/store/index.js
@@ -1,18 +1,15 @@
-import Vuex from 'vuex'
-import createPersistedState from 'vuex-persistedstate'
+import localStrage from 'localStrage'
 
-const initialState = {
+export const plugins = [ localStrage ]
 
+export const state = () => ({
   loggedIn: false
-}
 
-export const store = new Vuex.Store({
-  state: initialState,
-  mutations: {},
-  actions: {},
-  getters: {},
-  plugins: [createPersistedState()],
 })
+
+export const mutations = {
+  
+}
 
 export const actions = {
   nuxtClientInit({ commit }, context) {
